requirements.txt
aiosqlite==0.15.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
ecdsa==0.14.1
fastapi==0.61.1
h11==0.9.0
iso8601==0.1.13
passlib==1.7.2
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
PyPika==0.39.1
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.6
six==1.15.0
starlette==0.13.6
tortoise-orm==0.16.15
typing-extensions==3.7.4.3
uvicorn==0.11.8
websockets==8.1
wincertstore==0.2
